# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIH SaaS API — Contrat HTTP complet de la Phase 0 (SaaS Core) et des trois premiers increments de Phase 1
  version: 0.6.0
  description: >
    Premier fichier OpenAPI de ce depot (Phase 0, etape 5/13 : confirmation de paiement
    d'abonnement SaaS serveur-a-serveur, O-25.5). Etendu a l'etape 11/13 (ADR-0009) avec le
    PREMIER endpoint HTTP AUTHENTIFIE du depot : consultation du journal d'audit plateforme
    (`GET /api/v1/audit-entries`). Etendu a l'etape 12/13 item 11 (gap transverse, pas specifique
    a l'isolation/securite) pour couvrir les 10 routes reellement montees dans `createApp()`
    (`apps/api/src/server.ts`) : inscription/connexion/MFA (ADR-0010), recuperation break-glass
    `SUPER_ADMIN` (ADR-0005 Amendement 1) et `/health`. Etendu au premier increment vertical de
    Phase 1 (permission `tenant-config:administer`, decision de gouvernance 2026-09-28 sur le
    passage Phase 0 -> Phase 1) avec `GET`/`PATCH /api/v1/facility` : consultation et renommage de
    l'identite de l'etablissement du tenant courant (`HealthFacility`), PREMIER endpoint HTTP du
    module `tenant`. Etendu au deuxieme increment vertical de Phase 1 (permission
    `membership:administer`, meme decision de gouvernance) avec `GET /api/v1/memberships` :
    consultation paginee des memberships de l'etablissement du tenant courant, PREMIER endpoint
    HTTP de ce type pour le module `identity`. Document PUREMENT DESCRIPTIF : aucune
...
servers:
...
paths:
  /api/v1/payments/webhook:
    post:
      operationId: confirmSaaSPayment
      tags: [Payments]
      summary: Confirmation serveur-a-serveur d'un paiement d'abonnement SaaS (O-25.5)
      description: >
        Point d'entree UNIQUE de confirmation de paiement — jamais d'activation sur retour
        frontend (O-25.5). Signature HMAC obligatoire (en-tete `X-Payment-Signature`,
        `sha256=<hex>`) : toute notification absente de signature valide, illisible, ou portant
        un identifiant de transaction inconnu est rejetee SILENCIEUSEMENT (toujours HTTP 200,
        aucune information distinctive dans la reponse — evite qu'un tiers puisse sonder
        l'existence d'un `providerTransactionId` par la reponse HTTP). Idempotent par
        identifiant de transaction fournisseur : une meme notification recue plusieurs fois ne
        produit jamais d'effet applicatif supplementaire.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                Format PROPRE au prestataire (traduit en interne via un Anti-Corruption Layer,
                voir infrastructure/payment-provider/ — aucun prestataire reel choisi a ce
                stade, adaptateur SANDBOX documente dans le code). Exemple ci-dessous au format
                du sandbox actuellement cable.
              properties:
                transactionId:
                  type: string
                  example: sandbox_3fa85f64-5717-4562-b3fc-2c963f66afa6
                event:
                  type: string
                  enum: [payment.succeeded, payment.failed]
                  example: payment.succeeded
                occurredAt:
                  type: string
                  format: date-time
                  example: '2026-08-24T10:32:00Z'
      parameters:
        - name: X-Payment-Signature
          in: header
          required: true
          schema:
            type: string
            example: sha256=9f3a1c2e...
          description: Signature HMAC-SHA256 du corps brut exact de la requete.
      responses:
        '200':
          description: >
            Toujours renvoye, que la notification ait ete traitee, ignoree (deja traitee,
            idempotence) ou rejetee (signature absente/invalide, payload illisible, transaction
            inconnue) — reponse volontairement identique dans tous les cas (O-25.5, "rejet
            silencieux, pas d'erreur qui fuite d'info"). Corps toujours vide.
...
  /api/v1/registrations:
    post:
      operationId: registerHealthFacility
      tags: [Registration]
      summary: Inscription — creation d'un compte utilisateur ET d'un etablissement (ADR-0010 §2/§3)
      description: >
        Route NON authentifiee, CREATRICE D'ETAT DURABLE (ADR-0010 §Contexte 5) : le tenant cree
        n'est supprimable par AUCUN mecanisme du depot (O-03.1). Enchaine EXACTEMENT deux
        commandes dans la MEME requete serveur — `CreateUserAccount` PUIS
        `CreateHealthFacility(ownerUserId)`, `ownerUserId` etant TOUJOURS l'identifiant retourne
        par la premiere commande, jamais un champ du corps HTTP (anti mass-assignment, ADR-0008
        §9). `platformRole` est fixe a `NONE` cote serveur, jamais accepte du client. Si la
        seconde commande echoue apres la premiere (pathologique), le compte reste ORPHELIN,
        assume (aucune compensation, ADR-0008 §5) — le client recoit `500 internal_error`, sans
        distinction du cas nominal.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, facilityName]
              description: Schema `.strict()` (zod) — tout champ inconnu (`platformRole`, `tenantId`, `roleCodes`...) est rejete en `400`, jamais ignore silencieusement.
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 8
                  maxLength: 512
                facilityName:
                  type: string
                  minLength: 1
                  maxLength: 200
              example:
                email: admin@clinique-exemple.sn
                password: un-mot-de-passe-suffisamment-long
                facilityName: Clinique de l'Exemple
      responses:
        '202':
          description: >
            JAMAIS `201` : le tenant existe (`HealthFacility` `ACTIVE`) mais n'est pas encore
            ACCESSIBLE tant que la Saga de provisioning (Outbox) n'a pas complete
            `StartTrialSubscription` de maniere asynchrone (ADR-0008 §3). `status` est une
            constante litterale, jamais une lecture d'etat de Saga.
          content:
            application/json:
              schema:
                type: object
                required: [userAccountId, tenantId, status]
                properties:
                  userAccountId:
                    type: string
                    format: uuid
                  tenantId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [provisioning]
        '400':
          description: Corps malforme, champ inconnu (mass-assignment), ou JSON illisible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: invalid_request
        '409':
          description: L'email est deja associe a un `UserAccount` existant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: email_already_registered
        '429':
          description: >
            Limitation de debit (ADR-0010 §8) — cle par adresse IP (route anonyme, aucun sujet
            authentifie a ce stade). Valeurs NON DEFINITIVES (`RateLimitTuning.ts`).
          headers:
            Retry-After:
              description: Duree NOMINALE de la fenetre, en secondes entieres.
              schema:
                type: integer
              example: 60
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: too_many_requests
        '500':
          description: Erreur pathologique (compte cree, etablissement non cree — compte orphelin assume, ADR-0008 §5), ou toute autre defaillance technique.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: internal_error
  /api/v1/auth/sessions:
    post:
      operationId: createSession
      tags: [Sessions]
      summary: Connexion — authentification puis resolution de contexte (ADR-0010 §6)
      description: >
        Route NON authentifiee. Enchaine `AuthenticateUser` PUIS, seulement si l'identite est
        verifiee, `ResolveTenantContext` — jamais l'inverse. `context` est une SELECTION du
        client (jamais une preuve, revalidee integralement cote serveur contre les memberships
        reels). Derivation de l'intention en l'absence de `context` (ADR-0010 §6) : `SUPER_ADMIN`
        -> contexte `PLATFORM` ; exactement un tenant actif -> ce tenant ; zero ou plusieurs ->
        `200 { status: "context_selection_required" }`, aucune session ouverte. AUCUNE distinction
        email inconnu / mot de passe faux (anti-enumeration, un seul `401 invalid_credentials`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              description: Schema `.strict()`. DELIBEREMENT aucune contrainte de format sur `email`/`password` au-dela de la non-vacuite (ADR-0010 §6) — un format invalide produit `401`, jamais `400`, pour ne pas ouvrir un canal distinct exploitable par un attaquant.
              properties:
                email:
                  type: string
                password:
                  type: string
                context:
                  description: Selection EXPLICITE de contexte, optionnelle. Omise = derivation automatique (voir description).
                  oneOf:
                    - type: object
                      required: [kind, tenantId]
                      properties:
                        kind:
                          type: string
                          enum: [TENANT]
                        tenantId:
                          type: string
                          format: uuid
                    - type: object
                      required: [kind]
                      properties:
                        kind:
                          type: string
                          enum: [PLATFORM]
              example:
                email: admin@clinique-exemple.sn
                password: un-mot-de-passe-suffisamment-long
      responses:
        '200':
          description: >
            Trois formes mutuellement exclusives, discriminees par `status`. Sur `authenticated`,
            `Cache-Control: no-store` obligatoire (le corps porte `sessionId`/`refreshToken`
            en clair, ADR-0010 §10 regle 5).
          headers:
            Cache-Control:
              description: '`no-store` — present UNIQUEMENT sur la forme `authenticated` (secret en clair dans le corps).'
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthenticatedSessionResponse'
                  - $ref: '#/components/schemas/MfaRequiredResponse'
                  - $ref: '#/components/schemas/ContextSelectionRequiredResponse'
              examples:
                authenticated:
                  value:
                    status: authenticated
                    session:
                      sessionId: 01J4Z6X9K3Q7R8S9T0U1V2W3X4
                      kind: TENANT
                      tenantId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                      roleCodes: [ADMIN_ETABLISSEMENT]
                      permissionCodes: [audit:read]
                      absoluteExpiresAt: '2026-09-12T10:00:00.000Z'
                    refreshToken: 01J4Z6X9K3Q7R8S9T0U1V2W3X5.opaque
                mfaRequired:
                  value:
                    status: mfa_required
                    mfa:
                      pendingSessionId: 01J4Z6X9K3Q7R8S9T0U1V2W3X6
                      reason: CHALLENGE_REQUIRED
                      expiresAt: '2026-09-11T10:05:00.000Z'
                contextSelectionRequired:
                  value:
                    status: context_selection_required
                    availableTenantIds:
                      - 3fa85f64-5717-4562-b3fc-2c963f66afa6
                      - 9c858901-8a57-4791-81fe-4c455b099bc9
        '400':
          description: Corps malforme ou champ inconnu (mass-assignment).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: invalid_request
        '401':
          description: Identifiants invalides — indistinct entre email inconnu et mot de passe faux (anti-enumeration).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: invalid_credentials
        '429':
          description: Limitation de debit (ADR-0010 §8) — cle par adresse IP. Valeurs NON DEFINITIVES.
          headers:
            Retry-After:
              schema:
                type: integer
              example: 60
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: too_many_requests
        '500':
          description: Defaillance technique — jamais de detail interne expose.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: internal_error
  /api/v1/auth/sessions/mfa-challenge:
    post:
      operationId: verifyMfaChallenge
      tags: [Sessions, MFA]
      summary: Second facteur — verification du code MFA d'une session en attente (ADR-0010 §7 bis C)
      description: >
        `pendingSessionId` transporte EXCLUSIVEMENT en `Authorization: Bearer` (jamais le corps).
        NON montee derriere une authentification de contexte complet : une session `MFA_PENDING`
        serait structurellement refusee par ce middleware. Partage le meme presentateur de reponse
        que `POST /api/v1/auth/sessions` (branche `authenticated`) — jamais les trois autres
        formes possibles de cette derniere.
      security:
        - bearerSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [factor]
              description: Schema `.strict()`, union discriminee sur `factor.kind`.
              properties:
                factor:
                  oneOf:
                    - type: object
                      required: [kind, code]
                      properties:
                        kind:
                          type: string
                          enum: [TOTP]
                        code:
                          type: string
                          maxLength: 64
                    - type: object
                      required: [kind, code]
                      properties:
                        kind:
                          type: string
                          enum: [RECOVERY_CODE]
                        code:
                          type: string
                          maxLength: 64
              example:
                factor:
                  kind: TOTP
                  code: '123456'
      responses:
        '200':
          description: 'Session complete emise. `Cache-Control: no-store` obligatoire.'
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedSessionResponse'
        '400':
          description: Corps malforme ou champ inconnu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: invalid_request
        '401':
          description: >
            `Authorization` absent/malforme, `pendingSessionId` inconnu ou n'etant plus en
            attente de second facteur (reponse IDENTIQUE aux deux cas, anti-oracle d'etat de
            jeton), OU code TOTP/recuperation invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              examples:
                unauthenticated:
                  value: { error: unauthenticated }
                invalidCredentials:
                  value: { error: invalid_credentials }
        '403':
          description: Le contexte selectionne au moment de l'inscription n'est plus disponible (tenant suspendu/membership revoque entre-temps).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: forbidden
        '409':
          description: Aucun facteur MFA actif pour ce compte — l'enrolement doit etre complete d'abord.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: mfa_enrollment_required
        '429':
          description: >
            DEUX origines distinctes possibles, memes corps et code HTTP, `Retry-After`
            potentiellement different (assume par ADR-0010) : limitation de debit transport (cle
            IP, ADR-0010 §8) OU verrou anti-brute-force DE COMPTE (§7 bis D, duree NOMINALE du
            verrou).
          headers:
            Retry-After:
              schema:
                type: integer
              example: 60
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: too_many_requests
        '500':
          description: Defaillance technique, ou garde defensive interne (jamais atteinte en usage normal).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: internal_error
...
components:
  securitySchemes:
    bearerSessionAuth:
      type: http
      scheme: bearer
      description: >
        `sessionId` opaque (jamais un JWT auto-porteur) — ADR-0009 §8.3. Choix delibere et
        minimaliste, explicitement revisable a l'arrivee d'`apps/web` (transport definitif du
        `sessionId`, residu 5 de l'ADR).
  schemas:
...
    SimpleError:
      type: object
      description: >
        Format d'erreur produit par TOUTES les routes de ce contrat (`createErrorHandler`,
        `src/server.ts`, et chaque controleur individuellement) — PAS `ProblemDetails`
        (RFC 9457), jamais de detail interne expose. Ne pas confondre avec `ProblemDetails`
        ci-dessus, reserve aux futurs endpoints CRUD. L'enumeration ci-dessous est l'UNION de
        tous les codes reellement retournes par une route de ce fichier — une route donnee n'en
        retourne jamais qu'un sous-ensemble (voir ses reponses documentees).
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - unauthenticated
            - invalid_credentials
            - forbidden
            - mfa_required
            - mfa_enrollment_required
            - mfa_enrollment_already_active
            - email_already_registered
            - not_found
            - conflict
            - too_many_requests
            - internal_error
...
    AuthenticatedSessionResponse:
      type: object
      description: >
        Presentateur PARTAGE (ADR-0010 §7 bis C) entre `POST /api/v1/auth/sessions` (branche
        `authenticated`) et `POST /api/v1/auth/sessions/mfa-challenge` (seule reponse possible en
        succes) — `SessionResponseDto.ts`. `permissionCodes`/`roleCodes` exposes
        DELIBEREMENT (critere de sortie Phase 0 : "le frontend ne contient aucune permission
        codee en dur") ; `membershipId`/`userId`/`requiresMfa`/`mfaSatisfiedAt`/
        `sensitivityCategory`/`issuedAt` ne sont JAMAIS exposes.
      required: [status, session, refreshToken]
      properties:
        status:
          type: string
          enum: [authenticated]
        session:
          type: object
          required: [sessionId, kind, tenantId, roleCodes, permissionCodes, absoluteExpiresAt]
          properties:
            sessionId:
              type: string
            kind:
              type: string
              enum: [TENANT, PLATFORM]
            tenantId:
              type: [string, 'null']
              format: uuid
              description: '`null` pour un contexte `PLATFORM`.'
            roleCodes:
              type: array
              items:
                type: string
            permissionCodes:
              type: array
              items:
                type: string
            absoluteExpiresAt:
              type: string
              format: date-time
        refreshToken:
          type: [string, 'null']
          description: >
            Retourne UNE SEULE FOIS, jamais relisible (ADR-0006 §4) — `null` uniquement pour la
            reponse de `verifyMfaChallenge` dans le cas ou le contrat du handler le produit ainsi,
            jamais remplace par une chaine vide.
...